+++ src/fabric/GroupNativeComponent.js
@@ -25,7 +25,7 @@ type LayoutEvent = $ReadOnly<{|
 
 type SvgNodeCommonProps = $ReadOnly<{|
   name?: string,
-  opacity?: WithDefault<Float, 1.0>,
+  opacity?: WithDefault<Double, 1.0>,
   matrix?: $ReadOnlyArray<Float>,//CGAffineTransform,
   // transform?: ____TransformStyle_Internal, // CATransform3D, custom handling
   mask?: string,
